retyped functions to unsigned int and it seems that unsigned int and hexadecimals work now...

+++ src/ft_print_hex_low.c
@@ -6,7 +6,7 @@
 /*   By: smelicha <smelicha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/02 22:10:11 by smelicha          #+#    #+#             */
-/*   Updated: 2023/05/10 23:46:18 by smelicha         ###   ########.fr       */
+/*   Updated: 2023/05/18 19:45:11 by smelicha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int	ft_print_hex_low(t_data *data)
 	char	*string;
 	char	*ptr;
 
-	string = ft_hextoa(va_arg(*data->args, size_t), data);
+	string = ft_hextoa(va_arg(*data->args, unsigned int), data);
 	ptr = string;
 	data->varl = ft_strlen(string);
 	if (data->padnum && !data->dash)
